fix(frontend-flash): sign-extend AVM2 pushbyte operand
The swf crate stores PushByte values as u8, but AVM2's pushbyte
instruction sign-extends the byte to int. Cast through i8 first so
0xFF becomes -1 (not 255), fixing short-circuit evaluation in code
like `if (index < 0 || ...)` where the original AS3 relied on
negative sentinel values.

+++ crates/frontends/reincarnate-frontend-flash/src/translate.rs
@@ -564,7 +564,8 @@ fn translate_op(
         // Constants
         // ====================================================================
         Op::PushByte { value } => {
-            let v = fb.const_int(*value as i64);
+            // AVM2 pushbyte sign-extends: 0xFF is -1, not 255
+            let v = fb.const_int(*value as i8 as i64);
             stack.push(v);
         }
         Op::PushShort { value } => {
